requirements.txt
flask 
langchain 
openai 
pypdf 
tiktoken 
python-dotenv 
bs4 
langchain-openai 
langchain-community
langchain-pinecone
uvicorn 
gunicorn 
pinecone-client 
apscheduler
langchain-core 
langchain-experimental 
motor 
python-multipart 
pydantic[email]
pinecone 
boto3
pypdf2
pytest 
pyjwt 
python-jose
psycopg2
passlib